requirements.txt
discord.py==2.4.0
pynacl==1.5.0
ffmpeg==7.1.0
yt-dlp>=2025.6.30
python-dotenv>=1.0.1
elevenlabs==1.57.0
requests>=2.32.2
openai==1.63.0